+++ python_base_app/unix_user_handler.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2019  Marcus Rickert
+# Copyright (C) 2019-2024  Marcus Rickert
 #
 # See https://github.com/marcus67/python_base_app
 # This program is free software; you can redistribute it and/or modify
@@ -44,14 +44,26 @@ class UnixUserHandler(base_user_handler.BaseUserHandler):
 
         super().__init__(p_config=p_config, p_exclude_user_list=p_exclude_user_list)
         self._users = None
+        self._passwords = None
+
+        self._logger.warning(
+            "******************************************************************************************************")
+        self._logger.warning(
+            "Using the simple UnixUserHandler for user authentication which is inherently insecure.")
+        self._logger.warning(
+            "Consider upgrading to the more secure LdapUserHandler found in package python_base_app_ldap_extension.")
+        self._logger.warning(
+            "******************************************************************************************************")
 
         msg = "Using admin user '{username}'"
         self._logger.info(msg.format(username=self._config.admin_username))
 
         if self._config.user_list is not None:
             try:
-                self._users = {name_uid.split(":")[0].strip(): int(name_uid.split(":")[1]) for name_uid in
-                               self._config.user_list.split(",")}
+                entries = self._config.user_list.split(",")
+                self._users = {entry.split(":")[0].strip(): int(entry.split(":")[1]) for entry in entries}
+                self._passwords = {entry.split(":")[0].strip(): entry.split(":")[2] for entry in entries
+                                   if len(entry.split(":")) == 3}
 
             except Exception as e:
                 msg = "Invalid user:uid list '{list}' in [UnixUserHandler].user_list"
@@ -59,6 +71,10 @@ class UnixUserHandler(base_user_handler.BaseUserHandler):
 
             msg = "Using predefined users {users}"
             self._logger.info(msg.format(users=",".join(self._users.keys())))
+
+            if len(self._passwords) > 0:
+                self._logger.info(f"The following users have passwords defined: {','.join(self._passwords.keys())}")
+
 
     def list_users(self):
 
@@ -94,7 +110,8 @@ class UnixUserHandler(base_user_handler.BaseUserHandler):
 
     def authenticate(self, p_username, p_password):
 
-        return p_username == self._config.admin_username and p_password == self._config.admin_password
+        return ((p_username == self._config.admin_username and p_password == self._config.admin_password) or
+                p_username in self._passwords and self._passwords[p_username] == p_password)
 
     def is_admin(self, p_username):